Map file size bug fix & data update.
Fixed a bug where the maps were getting larger consecutively while the program ran - this was because the map never got 'closed' between saving, so all the data kept getting added to the file even though it wasn't displayed on the map. Very weird - I blame my usage of ipyleaflet to bulk save maps when it wasn't designed to do so. Future data updates should be smaller and faster to load!